pyproject.toml
[project]
name = "discord-claude"
version = "0.1.0"
description = "Discord bot that orchestrates multiple Claude Code CLI instances via tmux"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "python-dotenv>=1.0",
]

[project.scripts]
discord-claude = "discord_claude.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
